refactor: fully move docgen and component meta in renderer
