+++ code/frameworks/vue3-vite/package.json
@@ -51,14 +51,14 @@
   "dependencies": {
     "@storybook/builder-vite": "workspace:*",
     "@storybook/vue3": "workspace:*",
-    "magic-string": "^1.1.0",
-    "vue-component-meta": "^3.3.9",
-    "vue-docgen-api": "^4.75.1"
+    "magic-string": "^1.1.0"
   },
   "devDependencies": {
     "@types/node": "^22.19.1",
     "typescript": "^6.0.3",
-    "vite": "^7.0.4"
+    "vite": "^7.0.4",
+    "vue-component-meta": "^3.3.9",
+    "vue-docgen-api": "^4.75.1"
   },
   "peerDependencies": {
     "storybook": "workspace:^",
