+++ code/frameworks/vue3-vite/src/types.ts
@@ -4,15 +4,12 @@ import type {
 } from 'storybook/internal/types';
 
 import type { BuilderOptions, StorybookConfigVite } from '@storybook/builder-vite';
+import type { VueDocgenPlugin } from '@storybook/vue3';
 
-import type { ComponentMeta } from 'vue-component-meta';
-import type { ComponentDoc } from 'vue-docgen-api';
+export type { VueDocgenInfo, VueDocgenInfoEntry, VueDocgenPlugin } from '@storybook/vue3';
 
 type FrameworkName = CompatibleString<'@storybook/vue3-vite'>;
 type BuilderName = CompatibleString<'@storybook/builder-vite'>;
-
-/** Available docgen plugins for vue. */
-export type VueDocgenPlugin = 'vue-docgen-api' | 'vue-component-meta';
 
 export type FrameworkOptions = {
   builder?: BuilderOptions;
@@ -72,34 +69,3 @@ export type StorybookConfig = Omit<
 > &
   StorybookConfigVite &
   StorybookConfigFramework;
-
-/** Gets the type of a single array element. */
-type ArrayElement<T> = T extends readonly (infer A)[] ? A : never;
-
-/** Type of "__docgenInfo" depending on the used docgenPlugin. */
-export type VueDocgenInfo<T extends VueDocgenPlugin> = T extends 'vue-component-meta'
-  ? ComponentMeta
-  : ComponentDoc;
-
-/**
- * Single prop/event/slot/exposed entry of "__docgenInfo" depending on the used docgenPlugin.
- *
- * @example
- *
- * ```ts
- * type PropInfo = VueDocgenInfoEntry<'vue-component-meta', 'props'>;
- * ```
- */
-export type VueDocgenInfoEntry<
-  T extends VueDocgenPlugin,
-  TKey extends 'props' | 'events' | 'slots' | 'exposed' | 'expose' =
-    | 'props'
-    | 'events'
-    | 'slots'
-    | 'exposed'
-    | 'expose',
-> = ArrayElement<
-  T extends 'vue-component-meta'
-    ? VueDocgenInfo<'vue-component-meta'>[Exclude<TKey, 'expose'>]
-    : VueDocgenInfo<'vue-docgen-api'>[Exclude<TKey, 'exposed'>]
->;
