+++ code/renderers/vue3/src/docgen/vue-docgen-api.ts
@@ -1,0 +1,2 @@
+/** Legacy `vue-docgen-api` extraction, exposed through the docgen engine preset. */
+export { parse } from 'vue-docgen-api';
