+++ code/renderers/vue3/src/extractArgTypes.ts
@@ -7,7 +7,7 @@ import {
 } from 'storybook/internal/docs-tools';
 import type { SBType, StrictArgTypes, StrictInputType } from 'storybook/internal/types';
 
-import type { VueDocgenInfo, VueDocgenInfoEntry, VueDocgenPlugin } from '@storybook/vue3-vite';
+import type { VueDocgenInfo, VueDocgenInfoEntry, VueDocgenPlugin } from './types.ts';
 
 type PropertyMetaSchema = VueDocgenInfoEntry<'vue-component-meta', 'props'>['schema'];
