+++ code/renderers/vue3/src/preset.ts
@@ -25,4 +25,8 @@ export const experimental_importParsers = async (
   return [...input, vueImportParser];
 };
 
-export const experimental_vueDocgenEngine = async () => import('./docgen/component-meta.ts');
+/** Docgen extraction engines, keyed by plugin. Each loads lazily so a project only pays for the one it uses. */
+export const experimental_vueDocgenEngine = async () => ({
+  componentMeta: () => import('./docgen/component-meta.ts'),
+  vueDocgenApi: () => import('./docgen/vue-docgen-api.ts'),
+});
